Additional refactoring: reorganize hooks, move utility functions, clean up
