+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/index.ts
@@ -1,5 +1,1 @@
 export { DimensionPickerSidebar } from "./DimensionPickerSidebar";
-export {
-  DimensionPickerSidebarProvider,
-  useDimensionPickerSidebar,
-} from "./DimensionPickerSidebarContext";
