+++ frontend/src/metabase/metrics-viewer/hooks/index.ts
@@ -1,1 +1,1 @@
-export { useMetricsViewer } from "./use-metrics-viewer";
+export { useViewerState } from "./use-viewer-state";
