+++ frontend/src/metabase/metrics-viewer/test-utils.ts
@@ -1,22 +1,18 @@
-import type { UseMetricsViewerResult } from "metabase/metrics-viewer/hooks/use-metrics-viewer";
+import type { UseViewerStateResult } from "metabase/metrics-viewer/types";
 
 const EMPTY_AVAILABLE_DIMENSIONS = { shared: [], bySource: {} };
 
 export function createMockMetricsViewerResult(
-  overrides: Partial<UseMetricsViewerResult> = {},
-): UseMetricsViewerResult {
+  overrides: Partial<UseViewerStateResult> = {},
+): UseViewerStateResult {
   return {
     definitions: {},
     formulaEntities: [],
-    dimensionBreakouts: [],
     activeDimensionBreakout: null,
     initialLoadComplete: true,
-    loadingIds: new Set(),
-    resultsByEntityIndex: new Map(),
     queriesAreLoading: false,
     queriesError: null,
     modifiedDefinitionsBySlotIndex: new Map(),
-    breakoutValuesByEntityIndex: new Map(),
     metricSlots: [],
     series: [],
     cardIdToEntityIndex: {},
@@ -26,8 +22,10 @@ export function createMockMetricsViewerResult(
     sourceOrder: [],
     sourceDataById: {},
     availableDimensions: EMPTY_AVAILABLE_DIMENSIONS,
-    activeDimensionBreakoutAvailableDimensions: EMPTY_AVAILABLE_DIMENSIONS,
     sidebarAvailableDimensions: EMPTY_AVAILABLE_DIMENSIONS,
+    isSidebarOpen: true,
+    closeSidebar: jest.fn(),
+    openSidebar: jest.fn(),
     addMetric: jest.fn(),
     swapMetric: jest.fn(),
     removeMetric: jest.fn(),
